requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
authlib
requests
pydantic